Add reaction-based minigame
Minigame is played after the second voting screen.

Either blue or red bubbles spawn on the screen and the player must click the blue ones, avoiding the red ones.
If 10 blue bubbles are clicked, the player gets a point. If a red bubble is pressed, no points are awarded.
The screen transitions to the leaderboard after the minigame is over.