Add powerup screen (#132)
Sends a request to the backend to get a random powerup.
The screen then displays the powerup and a button to use it.
Upon pressing the button, a request is sent to the backend to carry out the logic of the powerup.
The screen then displays a message indicating that the powerup was used.